Function libmimalloc_sys::mi_realloc

pub unsafe extern "C" fn mi_realloc(
    p: *mut c_void, 
    newsize: usize
) -> *mut c_void

Re-allocate memory to newsize bytes.

Return pointer to the allocated memory or null if out of memory. If null is returned, the pointer p is not freed. Otherwise the original pointer is either freed or returned as the reallocated result (in case it fits in-place with the new size).

If p is null, it behaves as [mi_malloc]. If newsize is larger than the original size allocated for p, the bytes after size are uninitialized.