Function libmimalloc_sys::mi_realloc_aligned

pub unsafe extern "C" fn mi_realloc_aligned(
    p: *mut c_void, 
    newsize: usize, 
    alignment: usize
) -> *mut c_void

Re-allocate memory to newsize bytes, aligned by alignment.

Return pointer to the allocated memory or null if out of memory. If null is returned, the pointer p is not freed. Otherwise the original pointer is either freed or returned as the reallocated result (in case it fits in-place with the new size).

If p is null, it behaves as mi_malloc_aligned. If newsize is larger than the original size allocated for p, the bytes after size are uninitialized.